requirements.txt
aiogram==3.4.1
aiohttp-socks==0.9.0
python-dotenv==1.0.1
requests==2.32.3
pillow==10.4.0
fastapi==0.115.6
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
imap-tools==1.11.0
IMAPClient==3.0.1
httpx==0.27.2
